Workaround for TF/AWS bug: Ignore changes in sANs.
Terraform is perpetually recreating certain AWS ACM certificates because
the AWS API has started returning the list of subjectAltNames in a
nondeterministic order.

Work around this for the time being by ignoring changes to the
subjectAltNames. We will need to delete this line if we want to make
changes to the subjectAltNames in the future.

https://github.com/18F/identity-devops/issues/1469